feat: Move authentication logic to user edit page and clean up unused imports
